backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
geopy==2.4.1
requests==2.31.0
beautifulsoup4==4.12.2
pytest==7.4.3
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.1